# Conflicts:
#	utils.c

+++ common.h
@@ -75,7 +75,7 @@ typedef struct{
 
     char *filename; //TODO free memory after allocation
 
-    double **solution;
+    int **solution;
     int solution_size;
 
 } Tsp_prob;
